+++ app.rb
@@ -137,56 +137,6 @@ class App
   end
   # load books
 
-  def load_data
-    path = 'data/books.json'
-    if file_exist?(path)
-      books = JSON.parse(File.read('data/books.json'))
-      books.each do |book|
-        @books << Book.new(book['title'], book['author'])
-      end
-    else
-      puts 'books.json file does not exist'
-    end
-    # load people
-    path = 'data/people.json'
-    if file_exist?(path)
-      people = JSON.parse(File.read('data/people.json'))
-      people.each do |person|
-        case person['type']
-        when 'Student'
-          @people << Student.new(person['age'], person['classroom'], person['name'],
-                                 parent_permission: person['parent_permission'])
-        when 'Teacher'
-          @people << Teacher.new(person['age'], person['specialization'], person['name'])
-        end
-      end
-    else
-      puts 'people.json file does not exist'
-    end
-
-    # load rentals
-    path = 'data/rentals.json'
-    if file_exist?(path)
-      rentals = JSON.parse(File.read(path))
-      rentals.each do |rental|
-        @rentals << Rental.new(rental['date'], @books[rental['book']], @people[rental['person']])
-      end
-    else
-      puts 'rentals.json file does not exist'
-    end
-  end
-
-  def save_data
-    books = @books.map { |book| { title: book.title, author: book.author } }
-    File.write('data/books.json', JSON.pretty_generate(books))
-    rentals = @rentals.map do |rental|
-      { date: rental.date, book: @books.find_index(rental.book), person: @people.find_index(rental.person) }
-    end
-    File.write('data/rentals.json', JSON.pretty_generate(rentals))
-    people = @people.map { |person| { name: person.name, age: person.age, type: person.class } }
-    File.write('data/people.json', JSON.pretty_generate(people))
-  end
-
   def list_rentals_by_person_id
     puts # empty line
     print 'ID of person: '
