+++ data/rentals.json
@@ -1,7 +1,7 @@
 [
   {
-    "date": "2222-22-22",
-    "book": 0,
-    "person": 0
+    "date": "2222-01-01",
+    "book": 1,
+    "person": 1
   }
 ]
