+++ students.rb
@@ -2,7 +2,7 @@ require_relative 'persons'
 
 class Student < Person
   def initialize(age, classroom, name = 'Unknown', parent_permission: true)
-    super(age, name, parent_permission)
+    super(age, name, parent_permission: parent_permission)
     @classroom = classroom
   end
